[fix] move action command to Action folder
